set min amount value 100
