+++ src/init.rs
@@ -5,11 +5,8 @@ use crate::{
     },
     middleware::{ctx::Ctx, error::AppResult, mw_ctx::CtxState},
     routes::{
-        self, auth_routes,
-        community::profile_routes,
-        discussions, follows, notifications, posts, swagger, users,
-        wallet::{wallet_endowment_routes, wallet_routes},
-        webhooks::paypal,
+        self, auth_routes, community::profile_routes, discussions, follows, notifications, posts,
+        swagger, users, wallet, webhooks::paypal,
     },
     services::auth_service::{AuthRegisterInput, AuthService},
 };
@@ -145,13 +142,12 @@ pub async fn main_router(ctx_state: &Arc<CtxState>, wa_config: WebauthnConfig) -
         .merge(profile_routes::routes(ctx_state.upload_max_size_mb))
         .merge(tasks::routes())
         .merge(notifications::routes())
-        .merge(wallet_routes::routes())
-        .merge(wallet_endowment_routes::routes(ctx_state.is_development))
         .merge(users::routes(ctx_state.upload_max_size_mb))
         .merge(paypal::routes())
         .merge(discussions::routes(ctx_state.upload_max_size_mb))
         .merge(follows::routes())
         .merge(swagger::routes())
+        .merge(wallet::routes(ctx_state.is_development))
         .with_state(ctx_state.clone())
         .layer(CookieManagerLayer::new())
 }
