--- src/routes/wallet/mod.rs
@@ -1,3 +1,0 @@
-pub mod wallet_routes;
-// TODO should be private to this crate - or db service should be private so other crates can't call endowment method and change balance
-pub mod wallet_endowment_routes;
