+++ src/routes/webhooks/mod.rs
@@ -1,1 +1,2 @@
 pub mod paypal;
+pub mod stripe;
